ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
voyageai==0.3.2
pymongo==4.6.1
python-multipart==0.0.6
pymupdf==1.23.8
python-docx==1.1.0
requests==2.31.0
numpy==1.26.2
python-dotenv==1.0.0
sentence-transformers>=2.3.0

--extra-index-url https://download.pytorch.org/whl/cpu
torch

